FL/barcode_ocr_demo/requirements.txt
opencv-python
pillow
numpy
pandas
pyzbar
easyocr
pytesseract
python-barcode
fastapi
uvicorn
imutils
